requirements.txt
boto3==1.4.1
matplotlib>=2.0.0b4
numpy>=1.11.2
pandas>=0.19.1
paramiko>=2.0.2
apache-libcloud==1.5.0